Add JSDoc; consolidate redundant functions
